requirements.txt
mss
pygetwindow
pytesseract
Pillow
pandas
scikit-learn
joblib
streamlit
plotly
tensorflow
opencv-python
